pyproject.toml
[tool.poetry]
name = "taskmaster"
version = "0.5.1"
description = ""
authors = ["Andres Javier Lopez <code@andresjavierlopez.com>"]
readme = "README.md"
packages = [{include = "taskmaster"}]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.19"
asyncpg = "^0.28.0"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
greenlet = "^3.0.3"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
bcrypt = "4.0.1" # Avoid upgrading until passlib 1.7.5
python-multipart = "^0.0.9"
alembic = "^1.13.1"
authlib = "^1.3.1"
httpx = "^0.27.2"
langchain = "^0.3.3"
langchain-openai = "^0.2.2"
click = "^8.1.7"
twilio = "^9.2.2"
redis = "^5.0.7"
langgraph = "^0.2.38"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
# make it compatible with black
profile = "black"

[tool.pytest.ini_options]
addopts = "--cov=taskmasterexp"
testpaths = "taskmasterexp"
asyncio_mode = "auto"

[tool.coverage.run]
omit = [ "taskmasterexp/tests/*" ]
